Several major updates: 1) Users can now use the GUI to define completion behaviors. 2) The interface "OnCompletion" has been renamed to "CompletionBehavior". The old name was dumb (a known fact that problem to be fixed) and documentation kept referencing its implementations as completion behaviors anyway. 3) All getName methods were replaced by toString methods. The original justification for those methods having different implementations is irrelevant (and was weak anyway). 4) Created an enum inside Schedulr representing the user-definable behaviors (opening #19).
Supporting updates:
Schedulr now maintains class references for completion behaviors, so that new ones may be instantiated.
Enabled the ability to implement and extend in user-defined classes.
Added a line separator and the location of user-defined completion behaviors into Config.
Schedulr's getTasksMatchingRule method no longer throws an error when the rule is null.